requirements.txt
# CHRONOS Requirements
# For Streamlit Cloud deployment

# Core
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
networkx>=3.0
scikit-learn>=1.3.0

# PyTorch (CPU only for cloud deployment)
torch>=2.1.0
torch-geometric>=2.4.0

# Utilities
pyyaml>=6.0.1
tqdm>=4.66.0

# Baselines (optional)
lightgbm>=4.0.0
